+++ ValheimServersideQoL/Processors/PlayerSpawnedProcessor.cs
@@ -1,4 +1,7 @@
-﻿namespace Valheim.ServersideQoL.Processors;
+﻿using UnityEngine;
+using static Skills;
+
+namespace Valheim.ServersideQoL.Processors;
 
 sealed class PlayerSpawnedProcessor : Processor
 {
@@ -9,12 +12,16 @@ sealed class PlayerSpawnedProcessor : Processor
     readonly Dictionary<string, SpawnInfo> _spawnInfo = [];
     readonly Dictionary<int, List<ExtendedZDO>> _spawnedByPrefab = [];
     readonly Dictionary<ExtendedZDO, SpawnedState> _spawnedStates = [];
-    ExtendedZDO? _lastSummoningPlayer;
+    PlayerProcessor.IPeerInfo? _lastSummoningPlayer;
+    int _makeFriendlyChance;
+    int _levelUpChance;
+    int _tolerateLavaChance;
 
     sealed class SpawnedState
     {
         public ExtendedZDO? FollowTarget { get; set; }
-        public DateTimeOffset LastPatrolPointUpdate { get; set; }
+        public DateTimeOffset NextPatrolPointUpdate { get; set; }
+        public bool ChancesEvaluated { get; set; }
     }
 
     public override void Initialize(bool firstTime)
@@ -27,9 +34,15 @@ sealed class PlayerSpawnedProcessor : Processor
                 list.Clear();
         }
 
+        _makeFriendlyChance = Config.HostileSummons.MakeFriendly.Value ? 100 :
+            Math.Min(Config.Skills.BloodmagicMakeHostileSummonsFriendlyMinChance.Value, Config.Skills.BloodmagicMakeHostileSummonsFriendlyMaxChance.Value);
+
+        _levelUpChance = Math.Min(Config.Skills.BloodmagicSummonsMinLevelUpChance.Value, Config.Skills.BloodmagicSummonsMaxLevelUpChance.Value);
+        _tolerateLavaChance = Math.Min(Config.Skills.BloodmagicMakeSummonsTolerateLavaMinChance.Value, Config.Skills.BloodmagicMakeSummonsTolerateLavaMaxChance.Value);
+
         if (_spawnInfo.Count is 0)
         {
-            if (Config.HostileSummons.AllowReplacementSummon.Value || Config.HostileSummons.FollowSummoner.Value)
+            if (Config.HostileSummons.AllowReplacementSummon.Value || _makeFriendlyChance > -1)
             {
                 foreach (var item in ObjectDB.instance.m_items.Select(static x => x.GetComponent<ItemDrop>()))
                 {
@@ -40,8 +53,9 @@ sealed class PlayerSpawnedProcessor : Processor
                     Dictionary<int, List<ExtendedZDO>> dict = [];
                     foreach (var prefab in spawnAbility.m_spawnPrefab)
                     {
-                        if (prefab.GetComponent<Tameable>() is not null || prefab.GetComponent<Humanoid>() is not { m_faction: Character.Faction.PlayerSpawned })
+                        if (prefab.GetComponent<Humanoid>() is not { m_faction: Character.Faction.Players or Character.Faction.PlayerSpawned } humanoid)
                             continue;
+
                         var hash = prefab.name.GetStableHashCode();
                         if (!_spawnedByPrefab.TryGetValue(hash, out var list))
                             _spawnedByPrefab.Add(hash, list = []);
@@ -65,7 +79,7 @@ sealed class PlayerSpawnedProcessor : Processor
         }
 
         UpdateRpcSubscription("SetTrigger", OnZSyncAnimationSetTrigger,
-            Config.HostileSummons.AllowReplacementSummon.Value || Config.HostileSummons.FollowSummoner.Value);
+            Config.HostileSummons.AllowReplacementSummon.Value || _makeFriendlyChance > -1);
 
         if (!firstTime)
             return;
@@ -82,8 +96,42 @@ sealed class PlayerSpawnedProcessor : Processor
     /// <see cref="ZSyncAnimation.SetTrigger(string)"/>
     void OnZSyncAnimationSetTrigger(ZRoutedRpc.RoutedRPCData data, string name)
     {
-        if (!_spawnInfo.TryGetValue(name, out var spawnInfo) || !Instance<PlayerProcessor>().Players.TryGetValue(data.m_targetZDO, out _lastSummoningPlayer))
+        if (!_spawnInfo.TryGetValue(name, out var spawnInfo) || (_lastSummoningPlayer = Instance<PlayerProcessor>().GetPeerInfo(data.m_senderPeerID)) is null)
             return;
+
+        float? skill = null;
+        if (!Config.HostileSummons.MakeFriendly.Value && _makeFriendlyChance > -1)
+        {
+            if (skill is null)
+            {
+                skill = _lastSummoningPlayer.GetEstimatedSkillLevel(SkillType.BloodMagic);
+                if (float.IsNaN(skill.Value))
+                    skill = 0;
+            }
+            _makeFriendlyChance = Mathf.RoundToInt(Utils.Lerp(Config.Skills.BloodmagicMakeHostileSummonsFriendlyMinChance.Value, Config.Skills.BloodmagicMakeHostileSummonsFriendlyMaxChance.Value, skill.Value));
+        }
+
+        if (_levelUpChance > -1)
+        {
+            if (skill is null)
+            {
+                skill = _lastSummoningPlayer.GetEstimatedSkillLevel(SkillType.BloodMagic);
+                if (float.IsNaN(skill.Value))
+                    skill = 0;
+            }
+            _levelUpChance = Mathf.RoundToInt(Utils.Lerp(Config.Skills.BloodmagicSummonsMinLevelUpChance.Value, Config.Skills.BloodmagicSummonsMaxLevelUpChance.Value, skill.Value));
+        }
+
+        if (_tolerateLavaChance > -1)
+        {
+            if (skill is null)
+            {
+                skill = _lastSummoningPlayer.GetEstimatedSkillLevel(SkillType.BloodMagic);
+                if (float.IsNaN(skill.Value))
+                    skill = 0;
+            }
+            _tolerateLavaChance = Mathf.RoundToInt(Utils.Lerp(Config.Skills.BloodmagicMakeSummonsTolerateLavaMinChance.Value, Config.Skills.BloodmagicMakeSummonsTolerateLavaMaxChance.Value, skill.Value));
+        }
 
         if (!Config.HostileSummons.AllowReplacementSummon.Value)
             return;
@@ -94,9 +142,9 @@ sealed class PlayerSpawnedProcessor : Processor
                 continue;
 
             if (list[0].GetOwner() == data.m_senderPeerID &&
-                ZNetScene.InActiveArea(list[0].GetSector(), _lastSummoningPlayer.GetSector()))
+                ZNetScene.InActiveArea(list[0].GetSector(), _lastSummoningPlayer.PlayerZDO.GetSector()))
             {
-                RPC.Damage(list[0], new(float.MaxValue) { m_attacker = _lastSummoningPlayer.m_uid });
+                RPC.Damage(list[0], new(float.MaxValue) { m_attacker = _lastSummoningPlayer.PlayerZDO.m_uid });
             }
             else
             {
@@ -109,7 +157,7 @@ sealed class PlayerSpawnedProcessor : Processor
     protected override bool ProcessCore(ExtendedZDO zdo, IReadOnlyList<Peer> peers)
     {
         UnregisterZdoProcessor = true;
-        if (zdo.PrefabInfo.Humanoid is not { Humanoid.m_faction: Character.Faction.PlayerSpawned } || zdo.PrefabInfo.Tameable is { Tameable.m_startsTamed: true })
+        if (zdo.PrefabInfo.Humanoid is not { Humanoid.m_faction: Character.Faction.Players or Character.Faction.PlayerSpawned })
             return false;
 
         if (!_spawnedStates.TryGetValue(zdo, out var state))
@@ -117,7 +165,7 @@ sealed class PlayerSpawnedProcessor : Processor
             _spawnedStates.Add(zdo, state = new());
             zdo.Destroyed += x => _spawnedStates.Remove(x);
 
-            if (zdo.PrefabInfo.Tameable is null /*&& Config.Summons.AllowReplacementSummon.Value*/)
+            if (zdo.PrefabInfo.Tameable is null)
             {
                 if (!_spawnedByPrefab.TryGetValue(zdo.GetPrefab(), out var list))
                     _spawnedByPrefab.Add(zdo.GetPrefab(), list = []);
@@ -130,24 +178,58 @@ sealed class PlayerSpawnedProcessor : Processor
             }
         }
 
-        if (Config.HostileSummons.FollowSummoner.Value && state.FollowTarget is null)
+        if ((Config.HostileSummons.FollowSummoner.Value || _makeFriendlyChance > -1) && state.FollowTarget is null)
         {
             var playerName = zdo.Vars.GetFollow();
             if (!string.IsNullOrEmpty(playerName))
                 state.FollowTarget = Instance<PlayerProcessor>().Players.Values.FirstOrDefault(x => x.Vars.GetPlayerName() == playerName);
-            state.FollowTarget ??= _lastSummoningPlayer;
+            state.FollowTarget ??= _lastSummoningPlayer?.PlayerZDO;
             if (state.FollowTarget is not null && string.IsNullOrEmpty(playerName))
                 zdo.Vars.SetFollow(playerName = state.FollowTarget.Vars.GetPlayerName());
         }
 
-        if (Config.HostileSummons.MakeFriendly.Value && !zdo.Vars.GetTamed())
+        var tamed = zdo.Vars.GetTamed();
+
+        if (!state.ChancesEvaluated)
         {
-            UnregisterZdoProcessor = false;
-            RPC.SetTamed(zdo, true);
-            return true;
+            state.ChancesEvaluated = true;
+            var randomState = UnityEngine.Random.state;
+            UnityEngine.Random.InitState(zdo.Vars.GetSeed());
+            try
+            {
+                if (_makeFriendlyChance > -1 && !tamed)
+                {
+                    if (_makeFriendlyChance < 100 && UnityEngine.Random.Range(0, 100) > _makeFriendlyChance)
+                        return false;
+
+                    UnregisterZdoProcessor = false;
+                    RPC.SetTamed(zdo, true);
+                    zdo.Vars.SetTamed(true);
+                }
+
+                if (_levelUpChance > -1)
+                {
+                    var level = 1;
+                    while (level < Config.Skills.BloodmagicSummonsMaxLevel.Value && UnityEngine.Random.Range(0f, 100f) <= _levelUpChance)
+                        level++;
+                    if (level != zdo.Vars.GetLevel())
+                    {
+                        zdo.Vars.SetLevel(level);
+                        RecreateZdo = true;
+                    }
+                }
+
+                if (_tolerateLavaChance > -1 && zdo.PrefabInfo.Humanoid is { Humanoid.m_tolerateFire: false } &&
+                    zdo.Fields<Humanoid>().UpdateValue(static () => x => x.m_tolerateFire, UnityEngine.Random.Range(0, 100) <= _tolerateLavaChance))
+                {
+                    RecreateZdo = true;
+                }
+            }
+            finally { UnityEngine.Random.state = randomState; }
         }
 
-        if (Config.HostileSummons.FollowSummoner.Value && zdo.PrefabInfo.Tameable is null)
+        var follow = Config.HostileSummons.FollowSummoner.Value || (_makeFriendlyChance > -1 && tamed);
+        if (follow && zdo.PrefabInfo.Tameable is null)
         {
             var cfg = Config.Advanced.HostileSummons.FollowSummoners;
             var fields = zdo.Fields<MonsterAI>();
@@ -160,10 +242,10 @@ sealed class PlayerSpawnedProcessor : Processor
 
             UnregisterZdoProcessor = false;
             if (state.FollowTarget is not null &&
-                (DateTimeOffset.UtcNow - state.LastPatrolPointUpdate) > TimeSpan.FromSeconds(cfg.MoveInterval / 2) &&
+                DateTimeOffset.UtcNow is { } now && now > state.NextPatrolPointUpdate &&
                 Utils.DistanceXZ(zdo.GetPosition(), state.FollowTarget.GetPosition()) > cfg.MaxDistance / 2)
             {
-                state.LastPatrolPointUpdate = DateTimeOffset.UtcNow;
+                state.NextPatrolPointUpdate = now.AddSeconds(cfg.MoveInterval / 2);
                 var rev = zdo.DataRevision;
                 zdo.Vars.SetSpawnPoint(state.FollowTarget.GetPosition());
                 zdo.Vars.SetPatrol(true);